feat: Move window, tab, field methods to `WindowUtil` class.
