+++ src/main/java/org/spin/base/db/OrderByUtil.java
@@ -1,3 +1,17 @@
+/************************************************************************************
+ * Copyright (C) 2018-present E.R.P. Consultores y Asociados, C.A.                  *
+ * Contributor(s): Edwin Betancourt, EdwinBetanc0urt@outlook.com                    *
+ * This program is free software: you can redistribute it and/or modify             *
+ * it under the terms of the GNU General Public License as published by             *
+ * the Free Software Foundation, either version 2 of the License, or                *
+ * (at your option) any later version.                                              *
+ * This program is distributed in the hope that it will be useful,                  *
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of                   *
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                     *
+ * GNU General Public License for more details.                                     *
+ * You should have received a copy of the GNU General Public License                *
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.            *
+ ************************************************************************************/
 package org.spin.base.db;
 
 import org.adempiere.model.MBrowse;
@@ -6,6 +20,10 @@ import org.adempiere.model.MViewColumn;
 import org.compiere.util.Env;
 import org.spin.util.ASPUtil;
 
+/**
+ * Class for handle SQL Order By
+ * @author Edwin Betancourt, EdwinBetanc0urt@outlook.com, https://github.com/EdwinBetanc0urt
+ */
 public class OrderByUtil {
 
 	/**
@@ -28,7 +46,8 @@ public class OrderByUtil {
 	
 	/**
 	 * Get Order By Postirion for SB
-	 * @param BrowserField
+	 * @param browser
+	 * @param browserField
 	 * @return
 	 */
 	public static int getBrowserFieldOrderByPosition(MBrowse browser, MBrowseField browserField) {
